+++ packages/ketchup-lite/cypress/e2e/components/typewriter.cy.ts
@@ -67,6 +67,7 @@ describe("Methods", () => {
       kulPause: null,
       kulSpeed: null,
       kulStyle: null,
+      kulTag: null,
       kulValue: null,
     } as Required<KulTypewriterPropsInterface>);
   });