+++ packages/ketchup-lite/src/components/kul-showcase/kul-showcase.tsx
@@ -30,10 +30,29 @@ export class KulShowcase {
   //#endregion
 
   //#region Props
+  /**
+   * Customizes the style of the component. This property allows you to apply a custom CSS style to the component.
+   * @default ""
+   */
+  @Prop({ mutable: false }) kulScrollElement: HTMLElement = undefined;
+  /**
+   * Customizes the style of the component. This property allows you to apply a custom CSS style to the component.
+   * @default ""
+   */
   @Prop({ mutable: true, reflect: true }) kulStyle = "";
   //#endregion
 
   //#region Internal variables
+  #headers: { [K in KulShowcaseTitle]: HTMLDivElement | null } = {
+    Components: null,
+    Framework: null,
+    Utilities: null,
+  };
+  #sections: { [K in KulShowcaseTitle]: HTMLDivElement | null } = {
+    Components: null,
+    Framework: null,
+    Utilities: null,
+  };
   #kulManager = kulManagerInstance();
   //#endregion
 
@@ -42,7 +61,14 @@ export class KulShowcase {
     const currentValue = this.currentState[type];
 
     return (
-      <div class="section">
+      <div
+        class="section"
+        ref={(el) => {
+          if (el) {
+            this.#sections[type] = el;
+          }
+        }}
+      >
         {this.#prepHeader(type, currentValue)}
         <div class="flex-wrapper flex-wrapper--responsive">
           {currentValue ? this.#comps(type) : this.#cards(type)}
@@ -52,8 +78,15 @@ export class KulShowcase {
   }
   #prepHeader(type: KulShowcaseTitle, current: string): VNode {
     return (
-      <div class="header">
-        <h2>{current || type}</h2>
+      <div
+        class="header"
+        ref={(el) => {
+          if (el) {
+            this.#headers[type] = el;
+          }
+        }}
+      >
+        <kul-typewriter kulTag="h2" kulValue={current || type}></kul-typewriter>
         <div class={`navigation ${current ? "active" : ""}`}>
           <kul-button
             class="kul-full-height kul-full-width"
@@ -209,15 +242,22 @@ export class KulShowcase {
           kulData={cardDataset}
           kulSizeX="300px"
           kulSizeY="300px"
-          onKul-card-event={(e) => this.#handleCardClick(e, type)}
+          onKul-card-event={async (e) => {
+            const { eventType } = e.detail;
+
+            if (eventType === "click") {
+              await this.#handleCardClick(e, type);
+              this.#scrollToElement(type);
+            }
+          }}
         ></kul-card>
       );
     });
   }
-  #handleCardClick(
+  async #handleCardClick(
     e: CustomEvent<KulCardEventPayload>,
     type: KulShowcaseTitle,
-  ): void {
+  ) {
     if (e.detail.eventType === "click") {
       this.currentState = {
         ...this.currentState,
@@ -227,13 +267,35 @@ export class KulShowcase {
     }
   }
   #handleScroll = () => {
-    this.showScrollTop = window.scrollY > 300;
+    const scrollElement = this.kulScrollElement
+      ? this.kulScrollElement.scrollTop
+      : window.scrollY;
+    this.showScrollTop = scrollElement > 300;
   };
+  #handleScrollTop = async () => {
+    if (this.kulScrollElement) {
+      this.kulScrollElement.scrollTo({ top: 0, behavior: "smooth" });
+    } else {
+      window.scrollTo({ top: 0, behavior: "smooth" });
+    }
+  };
+  async #scrollToElement(type: KulShowcaseTitle) {
+    if (this.#sections[type]) {
+      this.#sections[type].scrollIntoView({
+        behavior: "smooth",
+        block: "start",
+      });
+    }
+  }
   //#endregion
 
   //#region Lifecycle hooks
-  connectedCallback() {
-    window.addEventListener("scroll", this.#handleScroll);
+  componentWillLoad() {
+    if (this.kulScrollElement) {
+      this.kulScrollElement.addEventListener("scroll", this.#handleScroll);
+    } else {
+      window.addEventListener("scroll", this.#handleScroll);
+    }
   }
   render() {
     return (
@@ -282,14 +344,18 @@ export class KulShowcase {
             kulIcon="chevron_up"
             kulStyling="floating"
             title="Scroll to top"
-            onClick={() => window.scrollTo({ top: 0, behavior: "smooth" })}
+            onClick={this.#handleScrollTop}
           ></kul-button>
         </div>
       </Host>
     );
   }
   disconnectedCallback() {
-    window.removeEventListener("scroll", this.#handleScroll);
+    if (this.kulScrollElement) {
+      this.kulScrollElement.removeEventListener("scroll", this.#handleScroll);
+    } else {
+      window.removeEventListener("scroll", this.#handleScroll);
+    }
   }
   //#endregion
 }