+++ packages/ketchup-lite-react/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-rc.4](https://github.com/lucafoscili/ketchup-lite/compare/2.4.0-rc.3...2.4.0-rc.4) (2024-12-07)
+
+**Note:** Version bump only for package ketchup-lite-react
+
+# [2.4.0-rc.3](https://github.com/lucafoscili/ketchup-lite/compare/2.4.0-rc.2...2.4.0-rc.3) (2024-12-07)
+
+**Note:** Version bump only for package ketchup-lite-react
+
+# [2.4.0-rc.2](https://github.com/lucafoscili/ketchup-lite/compare/2.4.0-rc.1...2.4.0-rc.2) (2024-12-07)
+
+**Note:** Version bump only for package ketchup-lite-react
+
+# [2.4.0-rc.1](https://github.com/lucafoscili/ketchup-lite/compare/2.4.0-rc.0...2.4.0-rc.1) (2024-12-07)
+
+**Note:** Version bump only for package ketchup-lite-react
+
+# [2.4.0-rc.0](https://github.com/lucafoscili/ketchup-lite/compare/2.3.3...2.4.0-rc.0) (2024-12-06)
+
+**Note:** Version bump only for package ketchup-lite-react
+
 ## [2.3.3](https://github.com/lucafoscili/ketchup-lite/compare/2.3.3-rc.0...2.3.3) (2024-12-06)
 
 **Note:** Version bump only for package ketchup-lite-react